+++ system-parachains/asset-hubs/asset-hub-kusama/tests/tests.rs
@@ -27,7 +27,8 @@ use asset_hub_kusama_runtime::{
 	AllPalletsWithoutSystem, AssetConversion, AssetDeposit, Assets, Balances, Block,
 	ExistentialDeposit, ForeignAssets, ForeignAssetsInstance, MetadataDepositBase,
 	MetadataDepositPerByte, ParachainSystem, PolkadotXcm, Runtime, RuntimeCall, RuntimeEvent,
-	RuntimeOrigin, SessionKeys, TrustBackedAssetsInstance, XcmpQueue, SLOT_DURATION,
+	RuntimeOrigin, SessionKeys, ToPolkadotXcmRouterInstance, TrustBackedAssetsInstance, XcmpQueue,
+	SLOT_DURATION,
 };
 use asset_test_utils::{
 	test_cases_over_bridge::TestBridgingConfig, CollatorSessionKey, CollatorSessionKeys,
@@ -585,6 +586,58 @@ fn reserve_transfer_native_asset_to_non_teleport_para_works() {
 			}
 		}),
 		WeightLimit::Unlimited,
+	);
+}
+
+#[test]
+fn report_bridge_status_from_xcm_bridge_router_for_polkadot_works() {
+	asset_test_utils::test_cases_over_bridge::report_bridge_status_from_xcm_bridge_router_works::<
+		Runtime,
+		AllPalletsWithoutSystem,
+		XcmConfig,
+		LocationToAccountId,
+		ToPolkadotXcmRouterInstance,
+	>(
+		collator_session_keys(),
+		bridging_to_asset_hub_polkadot,
+		|| bp_asset_hub_kusama::build_congestion_message(Default::default(), true).into(),
+		|| bp_asset_hub_kusama::build_congestion_message(Default::default(), false).into(),
+	)
+}
+
+#[test]
+fn test_report_bridge_status_call_compatibility() {
+	// if this test fails, make sure `bp_asset_hub_polkadot` has valid encoding
+	assert_eq!(
+		RuntimeCall::ToPolkadotXcmRouter(
+			pallet_xcm_bridge_hub_router::Call::report_bridge_status {
+				bridge_id: Default::default(),
+				is_congested: true,
+			}
+		)
+		.encode(),
+		bp_asset_hub_kusama::Call::ToPolkadotXcmRouter(
+			bp_asset_hub_kusama::XcmBridgeHubRouterCall::report_bridge_status {
+				bridge_id: Default::default(),
+				is_congested: true,
+			}
+		)
+		.encode()
+	)
+}
+
+#[test]
+fn check_sane_weight_report_bridge_status() {
+	use pallet_xcm_bridge_hub_router::WeightInfo;
+	let actual = <Runtime as pallet_xcm_bridge_hub_router::Config<
+		ToPolkadotXcmRouterInstance,
+	>>::WeightInfo::report_bridge_status();
+	let max_weight = bp_asset_hub_kusama::XcmBridgeHubRouterTransactCallMaxWeight::get();
+	assert!(
+		actual.all_lte(max_weight),
+		"max_weight: {:?} should be adjusted to actual {:?}",
+		max_weight,
+		actual
 	);
 }